% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2simi.R
\name{dist2simi}
\alias{dist2simi}
\title{Transform a distance matrix to a similarity matrix}
\usage{
dist2simi(D, method = "m")
}
\arguments{
\item{D}{A distance matrix.}

\item{method}{The method used for the transformation. It should
be \code{"m"}, \code{"e"}, \code{"es"}, \code{"r"}, or
\code{"rs"}. The default is \code{"m"}. See details.}
}
\value{
A similarity matrix.
}
\description{
Transform a distance matrix to a similarity matrix.
}
\details{
Available transformation methods are:
\tabular{llll}{
\code{"m"} \tab \tab \tab \cr
 \tab \tab \tab \eqn{s_{ij} = -0.5*d_{ij}^2} \cr
\code{"e"} \tab \tab \tab \cr   
 \tab \tab \tab \eqn{s_{ij} = exp(-0.5*d_{ij}^2)} \cr
\code{"es"} \tab \tab \tab \cr
 \tab \tab \tab   \eqn{s_{ij} = exp(-0.5*d_{ij}^2/mean(D^2))} \cr
\code{"r"} \tab \tab \tab \cr
 \tab \tab \tab    \eqn{s_{ij} = 1/(1+d_{ij})} \cr
\code{"rs"} \tab \tab \tab \cr
 \tab \tab \tab   \eqn{s_{ij} = 1/(1+d_{ij}/mean(D))} \cr
}
}
\examples{
library(MASS)
mu.x <- rep(2, 10)
sigma.x <- diag(rep(1,10))
n <- 30
X <- mvrnorm(n, mu.x, sigma.x)
B <- matrix(rnorm(10*100, mean = 3, sd = 3), nrow = 10)
Y <- X \%*\% B
D <- dist(Y)
S <- dist2simi(D, "m")
}
\author{
Jinjuan Wang.
}
