\name{drm-internal}
\alias{untangle.specials}
\alias{strata}
\alias{getpath.drm}
\alias{kronecker.drm}
\alias{kroneckerd.drm}
\alias{loglikd.drm}
\alias{logliks.drm}
\alias{na.include}


%- Also NEED an `\alias' for EACH other topic documented here.
\title{Internal drm functions}
\description{Internal drm functions}
\details{These are internal drm functions and are not to be called by
  the user.} 

\keyword{misc}%-- one or more ...
