% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.pf.R
\name{drm.pf}
\alias{drm.pf}
\title{Pseudo F test in distance regression}
\usage{
drm.pf(simi.mat, null.space, x.mat, permute = TRUE, n.monterCarlo = 1000,
  seed = NULL)
}
\arguments{
\item{simi.mat}{A similarity matrix among the subjects.}

\item{null.space}{A numeric vector containing the indices of those
columns in \code{x.mat} corresponding to the predictors which are
to be adjusted.}

\item{x.mat}{The matrix which combines the covariate matrix and
the matrix of interest.}

\item{permute}{A logical value indicating whether the
\emph{p}-value is calculated based on permutation procedure. The
default is \code{TRUE}.}

\item{n.monterCarlo}{The repeat number of the permutation
procedure. The default is \code{1000}.}

\item{seed}{The seed of the random number generator for the
permutation procedure. The default is \code{NULL}.}
}
\value{
The \emph{p}-value for the pseudo \emph{F} test statistic.
}
\description{
Calculate the \emph{p}-value for the pseudo \emph{F} test
statistic using permutation procedure.
}
\details{
x.mat[ ,null.space] indicates the covariate matrix and x.mat[
,-null.space] indicates the predictors of interest.
}
\examples{
library(MASS)
mu.x <- rep(2, 10)
sigma.x <- diag(rep(1,10))
B <- matrix(rnorm(10*100, mean = 3, sd = 3), nrow = 10)
n <- 30
X <- mvrnorm(n, mu.x, sigma.x)
Y <- X \%*\% B
S <- Y \%*\% t(Y)
drm.pf(S, 1:5, X)
}
\author{
Jinjuan Wang.
}
