% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.pf.R
\name{fast.pf}
\alias{fast.pf}
\title{Standard error difference in distance regression}
\usage{
fast.pf(X1, X2, S0, permutation = FALSE, perNum = 1000)
}
\arguments{
\item{X1}{A matrix containing the predictors to be adjusted.}

\item{X2}{A matrix containing the predictors of interest.}

\item{S0}{A similarity matrix.}

\item{permutation}{A logical value indicating whether the
permutation procedure is used. If \code{FALSE}, the \emph{p}-value
is calculated by the asymptotic distribution. The default is
\code{FALSE}.}

\item{perNum}{The repeat number of the permutation. The default is
\code{1000}.}
}
\value{
The \emph{p}-value for the sed test.
}
\description{
Calculate the \emph{p}-value for the standard error difference
(sed) test using the asymptotic distribution or the permutation
procedure.
}
\details{
If \code{permutation} is \code{FALSE}, the \emph{p}-value is
calculated using the standard normal distribution.
}
\examples{
library(MASS)
mu.x <- rep(2, 10)
sigma.x <- diag(rep(1,10))
B <- matrix(rnorm(10*100, mean = 3, sd = 3), nrow = 10)
n <- 30
X <- mvrnorm(n, mu.x, sigma.x)
Y <- X \%*\% B
S <- Y \%*\% t(Y)
X1 <- X[, 1:5]
X2 <- X[, 6:10]
fast.pf(X1, X2, S, permutation = FALSE)
}
\author{
Jinjuan Wang.
}
