\name{wheeze}
\alias{wheeze}
\docType{data}
\title{Wheeze among Steubenville children}
\usage{data(wheeze)}
\description{
  \code{wheeze} is a longitudinal dataset reporting the presence and
  absence of wheeze at ages 7, 8, 9 and 10 among 537 Steubenville children.  
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{id}{identification number of a child.}
    \item{wheeze}{a binary indicator of the presence of wheeze.}
    \item{age}{a numeric vector of age from 7 to 10.}
    \item{smoking}{a binary indicator of mother's smoking status at baseline.}
  }
}

\source{
Ekholm A, Smith PWF, McDonald  JW.  Marginal  regression
analysis  of  a  multivariate  binary response. \emph{Biometrika} 1995; 82(4):847-854.  
}

\references{Ekholm A, McDonald JW, Smith PWF.  Association  models
  for  a  multivariate  binary  response. \emph{Biometrics} 2000;
  56:712-718.}
\keyword{datasets}
